desktop: refactor to clean up desktop integration logic
